Fix lint and circular dep
