+++ packages/core/src/metrics/aggregator.ts
@@ -7,10 +7,11 @@ import type {
 import { timestampInSeconds } from '@sentry/utils';
 import type { BaseClient } from '../baseclient';
 import { DEFAULT_FLUSH_INTERVAL, MAX_WEIGHT, NAME_AND_TAG_KEY_NORMALIZATION_REGEX, SET_METRIC_TYPE } from './constants';
+import { captureAggregateMetrics } from './envelope';
 import { METRIC_MAP } from './instance';
 import { updateMetricSummaryOnActiveSpan } from './metric-summary';
 import type { MetricBucket, MetricType } from './types';
-import { captureAggregateMetrics, getBucketKey, sanitizeTags } from './utils';
+import { getBucketKey, sanitizeTags } from './utils';
 
 /**
  * A metrics aggregator that aggregates metrics in memory and flushes them periodically.
