+++ packages/core/src/metrics/browser-aggregator.ts
@@ -2,10 +2,11 @@ import type { ClientOptions, MeasurementUnit, MetricsAggregator, Primitive } fro
 import { timestampInSeconds } from '@sentry/utils';
 import type { BaseClient } from '../baseclient';
 import { DEFAULT_BROWSER_FLUSH_INTERVAL, NAME_AND_TAG_KEY_NORMALIZATION_REGEX, SET_METRIC_TYPE } from './constants';
+import { captureAggregateMetrics } from './envelope';
 import { METRIC_MAP } from './instance';
 import { updateMetricSummaryOnActiveSpan } from './metric-summary';
 import type { MetricBucket, MetricType } from './types';
-import { captureAggregateMetrics, getBucketKey, sanitizeTags } from './utils';
+import { getBucketKey, sanitizeTags } from './utils';
 
 /**
  * A simple metrics aggregator that aggregates metrics in memory and flushes them periodically.
